semantic-pattern-query-app/requirements.txt
# Semantic Pattern Query App - Production RAG Dependencies

# ============================================================================
# Vector Database
# ============================================================================
qdrant-client>=1.7.0

# ============================================================================
# Search Engine
# ============================================================================
elasticsearch>=8.0.0,<9.0.0  # Pin to 8.x for ES 8.x server compatibility

# ============================================================================
# Caching
# ============================================================================
redis>=5.0.0

# ============================================================================
# LLM and Embeddings
# ============================================================================
ollama>=0.1.0
sentence-transformers>=3.3.0
google-generativeai>=0.8.0  # Google Gemini API

# ============================================================================
# Document Processing
# ============================================================================
pypdf>=5.1.0
pdfplumber>=0.11.0
markdown>=3.5.0

# ============================================================================
# Numerical Operations
# ============================================================================
numpy>=2.0.0
scipy>=1.14.0

# ============================================================================
# API Framework
# ============================================================================
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0

# ============================================================================
# Utilities
# ============================================================================
python-dotenv>=1.0.0
tqdm>=4.67.0
click>=8.1.7

# ============================================================================
# Reranking
# ============================================================================
rank-bm25>=0.2.2

# ============================================================================
# Testing
# ============================================================================
pytest>=8.3.0
pytest-asyncio>=0.24.0

